Passing method getWordListForVerse() to containers to get flattened word list.
